BGDIINF_SB-1409: Moved class and functions out of tests/__init__.py
Based on the review moved the content of tests/__init__.py into
separates modules.
